move SumTree into per file
