+++ slm_lab/spec/base.json
@@ -167,9 +167,9 @@
       },
       "memory": {
         "name": "PrioritizedReplay",
-        "max_size": 4,
-        "e": 0,
-        "alpha": 1
+        "alpha": 1,
+        "epsilon": 0,
+        "max_size": 4
       },
       "net": {}
     }],
